fix Throwing pillow to not snap but still auto aim
Right joystick still rotates the player, but no longer snaps to a player.
Also, the auto-targeting is now used exlusively inside the Child's Throw
method. When throwing, it checks if there is a player in the direction the
player is pointing. If so, then the pillow is going this way, else it goes
in the Child's forward position.